pyproject.toml
[project]
name = "localrouter"
version = "0.2.8"
description = "Multi-provider LLM client with unified message format and tool support"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Center on Long-term Risk"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "anthropic>=0.50.0",
    "backoff>=2.2.1", 
    "openai>=1.98.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    "cache_on_disk>=0.4.0",
    "google-genai>=1.26.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pre-commit>=4.0.0",
    "black>=25.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["localrouter", "localrouter.*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pre-commit>=4.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
